pyproject.toml
[project]
name = "pythonalgunammda"
version = "0.1.0"
description = "MVP Sistema Courier Escalable"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.117.1",
    "sqlalchemy>=2.0.43",
    "pydantic>=2.0.0",
    "pydantic[email]>=2.0.0",
    "uuid>=1.30",
    "python-dotenv>=1.0.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "uvicorn[standard]>=0.22.0",
    "psycopg2-binary>=2.9.11",
    "dotenv>=0.9.9",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
